+++ source/renderer/app/utils/index.ts
@@ -1,7 +1,12 @@
 import { generateMnemonic } from './crypto';
+import { PersistentTimeMachine } from './PersistentTimeMachine';
+
+const timeMachine = new PersistentTimeMachine();
+timeMachine.init();
 
 export default {
   crypto: {
     generateMnemonic,
   },
+  timeMachine,
 };